Include the current panel.cloudatcost.com certificate and intermedate.
The way that the panel.cloudatcost.com SSL server has been configured is
not the best-practice - it omits sending the intermediate certificate that
is required to correctly validate the certificate chain.

I have worked around this locally by manually including the intermediate in
my local ca certificates, however for future reference the relevant cert
file is include in this commit.