Attempt to show really bad errors:
During an outage, the api was returning crazy data:

    > Host: panel.cloudatcost.com
    > User-Agent: curl/7.52.1
    > Accept: */*
    >
    < HTTP/1.1 200 OK
    < Date: Mon, 26 Jun 2017 09:36:55 GMT
    < Server: Apache/2.2.15 (CentOS)
    < X-Powered-By: PHP/5.3.3
    < Content-Length: 48
    < Connection: close
    < Content-Type: text/html; charset=UTF-8
    <
    * Curl_http_done: called premature == 0
    * Closing connection 0
    * TLSv1.2 (OUT), TLS alert, Client hello (1):
    DATABASE HAS LEFT THE BUILDING: CONNECTION_ERROR

This is an attempt to deliver the non-error http code, non-json error
message to the client in such a way that it can be displayed.

Of course, the outage was over before I tested my code changes.  But, with
the quality of these guys, I'm sure I'll get another try soon enough..